Fix error detection in VmScript's
When an artifact generator creates two artifacts using VmBuilds both use
the same temp-dir for the 'success' marker file which has the name
"SUCCESS" and is used to determine the success of a build.

That means that if the second vm script fails, the failure will not be
detected.

To fix that, change lxc execution to use different directories.