Make DSL/Interpreter more type safe
The 'addAction' function has the signature

   Handle -> IoCompiler () -> IoCompiler ()

where the second parameter is an action to execute after the actual
program was parsed and all the wirering and the build order was
determined. This allowes that action to update the state of IoCompiler,
e.g. by adding more (nested) actions. Since these actions will never be
executed due to the implementation details of 'compile' and because on a
high-level a phase seperation between wirering and IO actions is
generally useful, this commit changes the signature to

   Handle -> IoProgBuilder () -> IoCompiler ()

where IoProgBuilder has read-only access to the global state and hence
cannot add any  wirering or enque more 'IoProgBuilder's