Change the post-build script to run outside a container.
Allow the post build fix script in `VmImagesPostFix` to run on the build
machine instead of a container. This
allows the script to be used to fix for
example a boot loader installed in a
vm-image (in the VBR) after a resize
operation.

The script is called with the paths to all
raw vm image files in the isolated
build directory after the main build
script was executed and the images
were resized to their output size.